+++ src/Subscriptions/Iterators/AuthenticatingSyncIterator.php
@@ -25,10 +25,10 @@ class AuthenticatingSyncIterator implements SubscriptionIterator
         $previousGuardName = $this->configRepository->get('auth.defaults.guard');
 
         // Store the previous default Lighthouse guard name, so we can restore it after we're done
-        $defaultLighthouseGuardName = $this->configRepository->get('lighthouse.guard');
+        $defaultLighthouseGuardName = $this->configRepository->get('lighthouse.guards');
 
         // Set our subscription guard as the default guard for Lighthouse
-        $this->configRepository->set('lighthouse.guard', SubscriptionGuard::GUARD_NAME);
+        $this->configRepository->set('lighthouse.guards', [SubscriptionGuard::GUARD_NAME]);
 
         // Set our subscription guard as the default guard for the application
         $this->authFactory->shouldUse(SubscriptionGuard::GUARD_NAME);
@@ -59,7 +59,7 @@ class AuthenticatingSyncIterator implements SubscriptionIterator
             });
         } finally {
             // Restore the previous default Lighthouse guard name
-            $this->configRepository->set('lighthouse.guard', $defaultLighthouseGuardName);
+            $this->configRepository->set('lighthouse.guards', [$defaultLighthouseGuardName]);
 
             // Restore the previous default guard name
             $this->authFactory->shouldUse($previousGuardName);